Changes for properties file segregation

+++ controller/message.js
@@ -16,68 +16,85 @@ const logger = winston.createLogger({
 
 
 module.exports.newMessage = function(req, res){
-    try{
-        return res.render('message', {role:req.user.role});
-    }catch(err){
-        logger.error(err.toString())
-        return res.redirect('back');
+    if(req.user.role === 'Admin' || req.user.role === 'Teacher'){
+        try{
+            return res.render('message', {role:req.user.role});
+        }catch(err){
+            logger.error(err.toString())
+            return res.redirect('back');
+        }
+    }else{
+        return res.render('Error_403')
     }
 }
 
 module.exports.addMessageSchool = async function(req, res){
-    try{
-        await Messages.create({
-            Heading: req.body.Heading,
-            Message: req.body.Message,
-            SchoolCode: req.user.SchoolCode,
-            Category:req.body.Category,
-            Value:req.body.Value,
-        });
-        return res.status(200).json({
-            message:'Message sent'
-        })    
-    }catch(err){
-        console.log(err);
-        return res.status(500).json({
-            message:'Unable to send message'
-        })
+    if(req.user.role === 'Admin' || req.user.role === 'Teacher'){
+        try{
+            await Messages.create({
+                Heading: req.body.Heading,
+                Message: req.body.Message,
+                SchoolCode: req.user.SchoolCode,
+                Category:req.body.Category,
+                Value:req.body.Value,
+            });
+            return res.status(200).json({
+                message:'Message sent'
+            })    
+        }catch(err){
+            console.log(err);
+            return res.status(500).json({
+                message:'Unable to send message'
+            })
+        }
+    }else{
+        return res.render('Error_403')
     }
+    
 }
 
 
 module.exports.setNotificationsToClass = async function(req, res){
-    try{
-        await Messages.create({
-            Heading: req.body.Heading,
-            Message: req.body.Message,
-            SchoolCode: req.user.SchoolCode,
-            Category:'Class',
-            Value:req.params.Class,
-        })
-        return res.status(200).json({
-            message:'Notification sent'
-        })
-    }catch(err){
-        return res.status(500).json({
-            message:'Unable to create notification'
-        })
+    if(req.user.role === 'Admin' || req.user.role === 'Teacher'){
+        try{
+            await Messages.create({
+                Heading: req.body.Heading,
+                Message: req.body.Message,
+                SchoolCode: req.user.SchoolCode,
+                Category:'Class',
+                Value:req.params.Class,
+            })
+            return res.status(200).json({
+                message:'Notification sent'
+            })
+        }catch(err){
+            return res.status(500).json({
+                message:'Unable to create notification'
+            })
+        }
+    }else{
+        return res.render('Error_403')
     }
+    
 }
 
 
 module.exports.deleteMessage = async function(req, res){
-    try{
-        let message = await Messages.findByIdAndDelete(req.params.id);
-        return res.status(200).json({
-            message:'Message deleted successfully',
-            id:message._id
-        })
-    }catch(err){
-        return res.status(500).json({
-            message:'Unable to delete'
-        })
+    if(req.user.role === 'Admin' || req.user.role === 'Teacher'){
+        try{
+            let message = await Messages.findByIdAndDelete(req.params.id);
+            return res.status(200).json({
+                message:'Message deleted successfully',
+                id:message._id
+            })
+        }catch(err){
+            return res.status(500).json({
+                message:'Unable to delete'
+            })
+        }
+    }else{
+        return res.render('Error_403')
     }
-    
 }
 
 module.exports.getNotifications = async function(req, res){
@@ -115,15 +132,19 @@ module.exports.getNotifications = async function(req, res){
 }
 
 module.exports.deleteMessageSchool = async function(req, res){
-    try{
-        let message = await Messages.findByIdAndDelete(req.body.id);
-        return res.status(200).json({
-            message: 'Record deleted'
-        })
-    }catch(err){
-        console.log(err);
-        return res.status(500).json({
-            message:'Unable to delete'
-        })
+    if(req.user.role === 'Admin' || req.user.role === 'Teacher'){
+        try{
+            let message = await Messages.findByIdAndDelete(req.body.id);
+            return res.status(200).json({
+                message: 'Record deleted'
+            })
+        }catch(err){
+            console.log(err);
+            return res.status(500).json({
+                message:'Unable to delete'
+            })
+        }
+    }else{
+        return res.render('Error_403')
     }
 }
